fix: msrv and move datagram to exposed module
