+++ h3-quinn/Cargo.toml
@@ -15,7 +15,9 @@ license = "MIT"
 [dependencies]
 h3 = { version = "0.0.2", path = "../h3" }
 bytes = "1"
-quinn = { version = "0.10", default-features = false }
+quinn = { version = "0.10", default-features = false, features = [
+    "futures-io",
+] }
 quinn-proto = { version = "0.10", default-features = false }
 tokio-util = { version = "0.7.7" }
 futures = { version = "0.3.27" }
