+++ h3/src/proto/mod.rs
@@ -1,5 +1,4 @@
 pub mod coding;
-pub mod datagram;
 #[allow(dead_code)]
 pub mod frame;
 #[allow(dead_code)]
